// src/api/generated/models/ProductResultDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProductResultDto = {
  _id: string;
  productName: string;
  size: string;
  group: string;
  cost: number;
  sellingPrice: number;
  active: boolean;
};